requirements.txt
fastapi
uvicorn[standard]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic
pydantic-settings

